Return just the CourseDescriptor in discussions get_course()
Instead of calling the modulestore's get_course(), return the
descriptor via self.runtime.get_block(). Doing this means that bulk
ops caching works within CCX courses.

Note: The long term fix for this is to have some course-level scope
so that we're not peeking at the top of the tree just for some basic
discussion-specific configuration information (in this case the
blackout dates).

TNL-2697